+++ packages/gits_http/lib/gits_http.dart
@@ -9,3 +9,5 @@ export 'package:gits_http/src/usecase/use_case.dart';
 export 'package:gits_http/src/usecase/stream_use_case.dart';
 export 'package:gits_http/src/utils/either.dart';
 export 'package:gits_http/src/utils/gits_http_overrides.dart';
+export 'package:gits_http/src/utils/auth_token_option.dart';
+export 'package:gits_http/src/utils/refresh_token_option.dart';