+++ packages/gits_http/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+- add auth token option
+- add refresh token option
+- remove setToken from gits_http.dart
+
 ## 0.1.0
 
 - add documentation in readme.